Report
Analysis of Algorithms: Listed in the report is the known theoretical time complexities for finding the solution to the LCS(Longest Common Subsequence). I then took my implementation in Python and tested it for runtime performance with various input sizes. These values were recorded and used for asymptotic analysis. Results are listed in the conclusion and abstract of the report.